+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-20 03:07:10
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 6c94fcfe83d4eb507d895c722683b3ab4fb526e6
+
+## [ForwardWidgets-2kuai] - 2025-10-20 03:07:10
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: f86b7966ef7c03b5c2568ec10b2780b779d27631
+
+## [ForwardWidgets-pack1r] - 2025-10-20 03:07:10
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 59adb3a4ca57d051e82e363b9999ed707bbaaaa3
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-19 03:07:36
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main